+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Status: Available for use
 - Allow entrypoint suppression to be set in the configuration file - MINOR
 
 ### Fixed
+- Fix up old dependencies; move to thiserror/anyhow instead of failure - MINOR
 
 ## [0.6.2] - 2021-03-03